[feature/refactor-entities] Move to concerns, add timezone and checksum
