+++ tests/Entities/TestHelpers/CarbonFormattingEntityTest.php
@@ -19,3 +19,10 @@ it('can format carbon instances', function () {
         ->and($sut['dateOfBirth'])->toBe('01/01/1970')
         ->and($sut['someOtherDate'])->toBe('10 Dec 1963');
 });
+
+it('can provide time in the nominated timezone', function () {
+    $entity = new CarbonFormattingEntity();
+    $entity->timezoneDateTime = Carbon::create(2020, 1, 1, 12, 0, 0, '-7');
+
+    expect($entity->toArray()['timezoneDateTime'])->toBe('01/01/2020 19:00:00');
+});
