+++ tests/Entities/TestHelpers/ChecksumEntity2.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Midnite81\Core\Tests\Entities\TestHelpers;
+
+use Midnite81\Core\Entities\BaseEntity;
+
+class ChecksumEntity2 extends BaseEntity
+{
+    public int $id;
+
+    public string $name;
+}
